Inital Level Editor Created
Editor does a few different things:

Has a drop down menu to decide which element the player wants to create
Has a button that must be pressed in order for the elements to be drawn
Spawns can only be built on paths, same as obstacles (easy to change)
Only one player spawn can exist in the game at a time, multiple tank
spawns can exist (can be easily changed)
Has a clear button to completely reset the level to default.

Need to add functionality to spawn turtles (these turtles wouldn't have
any initial behaviour)
Potentially a limit should be set on the number of tank spawns in the
game.