Procedural Generation improvement
Progress has been made with the procedural generation.

Now, each patch (based on where it is in the game) is defined as
belonging to a tile. Now, when the random button is pressed, each patch
is asked where it exists in the game and, based on its position,
assigned a tileNumber. Then, based upon which tileNumber the patch has,
it is then assigned a colour. This is also randomised, so that upon
pressing the button, the colours of the "tiles" change each time.

Next thing that needs to happen is each tile is actually an image that
we have defined, that loads in. Then this section only needs bug tested.