Bug fixing
Fixed bug that stopped it from cycling through all three of the tiles so
far, only used first two.

Noticed bug where sometimes NO tile is loading, and other times two
tiles load on top of each other. Will investigate this when more tiles
are defined to see if lack of defined tiles is the issue.