Changed some structure, see description
Seem to have fixed bug of tiles loading on top of each other by
separating the long bunch of IF statements that, based upon what the
value of the chosenTile variable was, loaded the corresponding tile,
into its own method.

This method will work for all of the 9 chosenTile variables, as they are
simply passed into it as a parameter. Also takes in the x and y offsets
as well.

Changed code structure so that the generate-map method is above all of
the methods which it calls. Doesn't make a difference to execution,
improves readability.