Added in fuel and ammo spawns
Whilst this will obviously be turtles in the final version, I've quickly
created the ammo and fuel spawning at random by changing the colours of
patches, just to demonstrate this.

Next thing to do is to create about 7-8 more tiles, change the ammo and
fuel spawns to turtles, and then integrate into full extended project
and bug test.

HUZZAH!