Changed how tiles are made, see description.
Before, each tile had multiple if statements, which was a lot of code
repeating itself, with slightly different values being checked. Now, a
method exists to create each obstacle where it needs to be for each
tile, making this process much more readable and saving space in the
code. Probably won't affect efficiency of the code.