Read description
Fixed issue with bombs spawning on top of walls/obstacles. (Also
refactored code to do this).
Fixed issue with fuel pickups spawning on top of walls/obstacles. (Also
refactored code to do this).
Fixed issue which allowed for playerSpawn to be placed on top of
wall/obstacle in the level editor.